pyproject.toml
[tool.poetry]
name = "CocinaRandomBackend"
version = "0.1.0"
description = "Backend para App de cocina"
authors = ["Vicente Espinosa"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
djongo = "^1.3.6"
python-dotenv = "^0.19.0"
pymongo = {extras = ["srv"], version = "^3.12.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
